requirements.txt
pandas~=2.2.3
python-dotenv~=1.1.0
openai~=1.82.0
requests~=2.32.0
bert-score~=0.3.13
nltk~=3.9.1
nlpaug~=1.1.11
rouge_score~=0.1.2
transformers~=4.52.4